refactor versus to FSD
