+++ src/fsd/0-app/routing/desktop-routes.tsx
@@ -1,8 +1,7 @@
 ﻿import React from 'react';
 import { RouteObject } from 'react-router-dom';
 
-import { versusLazyRoute } from '@/v2/features/vs/versus.route';
-
+import { versusLazyRoute } from '@/fsd/1-pages/learn-versus';
 import { guildApiLazyRoute } from '@/v2/pages/guild-api/guild-api.route';
 import { campaignProgressionLazyRoute } from 'src/v2/pages/campaign-progression/campaign-progression.route';
 import { dirtyDozenLazyRoute } from 'src/v2/pages/dirty-dozen/dirty-dozen.route';
