+++ src/fsd/1-pages/learn-versus/damage-chart.tsx
