+++ src/fsd/1-pages/learn-versus/equipment-select.tsx
