+++ src/fsd/1-pages/learn-versus/index.ts
@@ -1,0 +1,1 @@
+export { versusLazyRoute } from './versus.route';
