+++ src/fsd/1-pages/learn-versus/versus.tsx
